Support dynamic quantum architecture for IQM servers
For IQM servers fetch the quantum architecture from server, derive the dynamic architecture from this and write it to an architecture file. This file is then passed in the lowering pipeline in place of the previously used static architecture files.
Removed parameter "--iqm-machine" which was used to specify the IQM quantum architecture at compile time.
Removed static architecture files provided for Adonis, Apollo, and Aphrodite architectures.
Removed machine architecture bindings from IQM tests.
Implemented getting the authorization token from IQM_TOKEN environment variable. If this is not set the previous mechanism reading the token from a JSON file is used.
The output from the algorithm for finding usable qubits generates a list of qubits which are calibrated for mz, cz and prx. A mapping ensures that qubit names are enumerated in a linear manner and not fully calibrated qubits do not show.
[SW-1349]

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>