Fetch the dynamic QA only once
The ServerHelper class is instantiated twice and the initialize() is
called every time. The 2nd instance is needed to get the "future"
resonse and does not need the dynamic QA. Moved the fetching of the
dynamic QA from initialize() to updatePassPipeline() in which the file
with the target quantum architecture is needed.