Change algorithm finding usable qubits
The algorithm for finding usable qubits was changed. As result qubits
without the required connectivity are not listed in the output anymore.
Also the order has changed from exponential to linear.
A mapping of qubit names was added to translate the physical qubit names
used by CUDA-Q to the physical qubit names used by the IQM backend.