Make quantum architecture file temporary
Create a unique filename for the dynamic quantum architecture file and open it in the system temp directory.
Inject the generated filename on the pass pipeline.
Remove the dynamic quantum architecture file when done.
Cleanup: move the init function out of the class declaration.
Cleanup: improved doxygen headers and few other comments.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>