Added mapping of qubits on the mock server
Added support for mapping of qubits syntax in the job request to the mock server.
Deliberately removed QB2 and QB2 from the list of PRX calibrated gates to force the CUDA-Q code t work with an imperfect calibrated system. This results in changes in the topology and requries the mock server to support mapping of qubits.
Implemented a full checking of gates used in circuits versus the topology to verify that all 2 qubit gates are supported by the mock qpu.

Signed-off-by: Bernd Hoffmann <bernd.hoffmann@meetiqm.com>